PJAS-1670 Error handling on FE, vhost support in BE
