+++ Dockerfile
@@ -12,5 +12,6 @@ RUN npm install
 RUN npm run build
 
 EXPOSE 8080
+EXPOSE 4433
 
 CMD ["npm", "run", "start-debug"]
