+++ docker-compose.yml
@@ -4,7 +4,8 @@ services:
     image: metallic
     build: .
     ports:
-      - "8080:8080" # serve port
+      - "8080:8080"
+      - "4433:4433"
       - "9229:9229" # debug port
     networks:
       - jpps
