+++ package.json
@@ -26,6 +26,7 @@
     "react-tsparticles": "^2.9.3",
     "redis": "^4.6.13",
     "tsparticles": "^2.9.3",
+    "vhost": "^3.0.2",
     "vite-plugin-rewrite-all": "^1.0.1",
     "vite-plugin-svgr": "^2.4.0",
     "winston": "^3.11.0"
