+++ sslcert/lead-jobs.pjas.ch+1.pem
@@ -1,0 +1,27 @@
+-----BEGIN CERTIFICATE-----
+MIIEnDCCAwSgAwIBAgIQOvMXniPRw7DJ8DexSdlN+DANBgkqhkiG9w0BAQsFADCB
+qzEeMBwGA1UEChMVbWtjZXJ0IGRldmVsb3BtZW50IENBMUAwPgYDVQQLDDdkZ2pv
+cmdqaWV2QERyYWdhbnMtTWFjQm9vay1Qcm8ubG9jYWwgKERyYWdhbiBHam9yZ2pp
+ZXYpMUcwRQYDVQQDDD5ta2NlcnQgZGdqb3JnamlldkBEcmFnYW5zLU1hY0Jvb2st
+UHJvLmxvY2FsIChEcmFnYW4gR2pvcmdqaWV2KTAeFw0yNDAyMDkxMjMxNTBaFw0y
+NjA1MDkxMTMxNTBaMGsxJzAlBgNVBAoTHm1rY2VydCBkZXZlbG9wbWVudCBjZXJ0
+aWZpY2F0ZTFAMD4GA1UECww3ZGdqb3JnamlldkBEcmFnYW5zLU1hY0Jvb2stUHJv
+LmxvY2FsIChEcmFnYW4gR2pvcmdqaWV2KTCCASIwDQYJKoZIhvcNAQEBBQADggEP
+ADCCAQoCggEBANQZhX+jnoQmSYkm+i4UIdlfS2CYAiLRpTQj+n+l2/EeunBwKUGZ
+JJh7MHXAT+UqudRwuuVctsSc+g9oHwDl3e+5PnZLAPHiAT6DSvCqeVkWEmbMKYGF
+2x3OXv8pq0aerc7ApXi37XeCb2mbVgBTrJ7gC64O//Azy6SWC/mMBHl3UZuYqE3t
+ZhupSoYgw6COEoGJvlbi9nNEYYx5muzD5SixuaHSpKJbqLoDQL+0znrX+cE0i5MG
+E5jmqm4e92SGetW3OkGesjBsWD67KT0yf3eIUSmTksAlqk/rnE7B1CTIdNdTCsiS
+T2pJnBHuvEKsU1ApfA5L2eXC5zQCxof6X0sCAwEAAaN7MHkwDgYDVR0PAQH/BAQD
+AgWgMBMGA1UdJQQMMAoGCCsGAQUFBwMBMB8GA1UdIwQYMBaAFBO0PXuFPRzeJDz4
+Nh7+ZLqjrq2OMDEGA1UdEQQqMCiCEWxlYWQtam9icy5wamFzLmNoghMqLmxlYWQt
+am9icy5wamFzLmNoMA0GCSqGSIb3DQEBCwUAA4IBgQDZyMqC5rYwadkgO/RW1Ja1
+Ch78Mbpe6URe95pe1zhyLi2+g3zJ7w7eoO9vPE6kc7kh3VKiHlGC5bFZ7nnn5Smo
+i1IMb/qgyyXt6d3obPcbtII+5r1dGyWcTnS9XJnlAxibJUR5/yf5lYPyxxDUfysJ
+WNf9v47e/foUYfB7fZNmNVGgjfcWHtRAlURSQcfyNaF9GTC5XnDmumY1Y6bmrFD0
+zlreBCXKfrItWBv8wz70kIwCJAH0p2uYDPARK7XgZrObdae5j2piqbKFNJ1UZ71u
+0SMltRFKyQ1TqUtnHj9+gqZzk2H7K8dKKCxRMb281YnwXp0pTqGYtAPvYiawnMZd
+7PVoKTA2CCQ18bfrx8RGsdKVmSNCnBiCfrIQovaMIjOg0pTH3hMDAgkjqeH3AbEg
+o3JGy6TOygRzCvw5reQCTtar+QgTDlG40A6EyGwavMK4emetj/pJtZ1zlZnqkbm/
+hKkjQjAPfbnK48gpKGvxF1PlK6ieQqtY04FsuiKF9ro=
+-----END CERTIFICATE-----
